changed input to int.

+++ src/SphereCircumference.java
@@ -3,9 +3,9 @@ public class SphereCircumference implements ISubscriber {
     public void notifySubscriber(String input) {
         System.out.println("Hello, I am Sphere Circumference and I am notified with " + input);
 
-        double radius;
+        int radius;
         try {
-            radius = Double.parseDouble(input);
+            radius = Integer.parseInt(input);
             if (radius < 0) {
                 System.out.println("Radius is negative.");
                 return;
@@ -19,7 +19,7 @@ public class SphereCircumference implements ISubscriber {
         System.out.println("Radius is " + radius + ", Sphere Circumference is: " + doCalculation(radius));
     }
 
-    public double doCalculation(double input) {
+    public double doCalculation(int input) {
         return 2 * Math.PI * input;
     }
 }
